refact(boilerplate): rename files
